+++ src/form-elements/FormElementTelephone.tsx
@@ -84,6 +84,7 @@ function FormElementTelephone({
             value={value}
             validationMessage={validationMessage}
             lookupButtonConfig={element.lookupButton}
+            overrideRequiredMessage={undefined}
           />
         </div>