+++ src/form-elements/FormElementText.tsx
@@ -80,6 +80,7 @@ function FormElementText({
             value={value}
             validationMessage={validationMessage}
             lookupButtonConfig={element.lookupButton}
+            overrideRequiredMessage={undefined}
           />
         </div>
         {(isDisplayingValidationMessage || !!element.maxLength) && (