+++ src/form-elements/FormElementTime.tsx
@@ -109,6 +109,7 @@ function FormElementTime({
             value={value}
             validationMessage={validationMessage}
             lookupButtonConfig={element.lookupButton}
+            overrideRequiredMessage={undefined}
           />
         </div>