+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `formsAppService.getFormSubmissionMetaList()`
 - `hideEndAdornment` to `AutocompleteDropdown`
 
+### Changed
+
+- required logic for `lookupButton` form element
+
 ## [10.1.2] - 2026-05-07
 
 ### Added